Start work to support poetry
In addition to conda.

Note that, currently, pyproject.toml isn't really generated by
poetry and doesn't list dependencies, so "poetry install" in the
onCreate script would not have the desired effect. That command is
therefore commented out for now, and a possible procedure for
making and using a pyproject.toml file with poetry-managed
dependencies is listed in the revised comment in pyproject.toml.

This retains full conda support, which should not be removed.
However, the base environment is no longer activated automatically.
This is to help avoid accidentally *mixing* conda and poetry
environments in the same running shell at the same time, which
behaves unexpectedly. For example, poetry commands shouldn't use
python interpreters in conda environments, not even the base
environment. It remains generally unnecessary to activate the base
environment, because the conda and mamba commands remain available,
even when no environment is activated. They should not be run when
in a poetry-managed virtual environment or shell.