Let CI test setup steps run with full concurrency
The approach in the previous commit of limiting jobs took way too
long because most of the time was provisioning micromamba. Caching
to GitHub Actions storage wouldn't be a very good solution, because
the cached data would be different (and thus shouldn't be shared)
for each of the fifteen jobs, and because we would want a very
short cache window (with conda, at least currently, we are not
pinning specific dependency versions, and CI should reflect that).

This new approach attempts to lock just the "Run Unit Tests" step
using a mutex. It should let just one of these steps run at a time,
across all jobs, while still allowing each job to provision with
micromamba concurrently. However, this may not be worth it, even if
it works, because of the complexity entailed in using an additional
branch of the repository to maintain the mutex state, as the mutex
action being used does.