It helps if I check out the repo first
This problem was not related to paths.

Since I have to check out the repository, the space and code
complexity savings in test.yml are not as pronounced. But it is
still shorter and simpler. Maybe more important is how having the
logic in a script lets shellcheck scan it.