Add single CI job to run test modules as scripts
This is a non-matrix CI job that runs only on Ubuntu with the
version of Python called for in environment.yml (currently Python
3.11). It runs the test modules as scripts in a shell loop, exiting
with a successful status if all tests passed (besides skipped
ones), or a failure status if any failed. Even if tests fail, the
rest of the test modules are still run as scripts before the job
terminates. This is done in a single step of the job for all the
scripts, which are detected by path (their names are not hard-coded
into the workflow).

In-memory caching of embeddings is not performed in this test job.

This serves two purposes:

- It ensures all the test modules can be run as scripts. If any are
  not marked executable or don't have a working hashbang and logic
  to run the tests when the module is run as a script, this will
  fail or at least demonstrate the problem when the CI output is
  cursorily examined (e.g., for reported number of tests run per
  module). Likewise, if test code or test helper code contains a
  bug that keep the tests from working when run as scripts, this
  will smoke that out (unless the bugs make tests wrongly pass).

- This runs all the tests on one platform with one Python version,
  without any in-memory caching. We are still saving API calls,
  even compared to not doing in-memory caching on CI but omitting
  this new test job, because in-memory caching currently decreases
  the number of API calls by almost half, and there are fifteen
  other test jobs currently (the five supported Python versions and
  the three major platforms).

It makes sense to take this opportunity to have a single CI job
that runs all tests without in-memory caching of embeddings,
because that makes conceptual sense and is a fairly strong design
choice for a script-running test job anyway: separate runs of
scripts wouldn't share in-memory cached embeddings with each other
anyway. (That might currently be a minor or irrelevant limitation,
but in terms of scalability the benefit would still be less.)

All the old test jobs are of course retained unchanged. This adds
to them; it does not replace any of them.