Rename "pairs" to "whatalgo"; drop A/B/C from docstrings
Removing the distinguishing information from the module docstrings
of the input code makes it so a model doesn't get let to (or away
from) a correct answer by guesses based on it. For example, if way
B happens to be BFS, are we inadvertently helping the model figure
that out?