Regenerate whatalgo embeddings
To save them in as safetensors instead of JSON.

This doesn't fix anything, it just reruns the notebook to save the
embeddings in the form that I'm probably going to want to use for
other texts' embeddings in whatalgo (such as descriptions generated
from algorithm and data structure names, which I hope will compare
more similar to the code examples they relate to than the names).
This does not create or embed any such other texts yet.