LLM prompts from short list of names
This is for semi-manual experiments, and if I do experiments with
lower-availability or more expensive models.

I plan to embed all ~940 of the names from the full list using
such prompts, or a modified prompt if experiments with these
suggest it is not the best way to prompt for a short and accurate
definition free of extraneous details. But for that the prompts
will be automatically constructed.