Improve style; clarify summary
In regard specifically to the change of one import from relative to
absolute, the reason is not that I consider relative imports to be
a worse style in this case (where it is importing a submodule).
Rather, isort separates the import into its own group when it is
relative, in a way that I don't think improves clarity; making it
absolute avoids that problem without making isort's configuration
more complex or adding a suppression.

That is only a small part of the changes in this commit, but I
think the other changes are largely self-explanatory.