Document _NAME_PATTERN
This turns whatalgo._NAME_PATTERN into a verbose regex with
comments that explain each part in a way that makes their purposes
fairly apparent.

This is mostly a refactoring of the pattern. It should have the
same behavior as before, except that it is more Unicode-aware.
Regenerating the name list with it does not produce a difference.

This is not a simplification, but I've removed the FIXME saying to
simplify it, because I haven't thought of a good way to simplify
it, and documenting it this way makes it clear.