Use orjson for saving/loading algorithm names
Instead of the standard-library json module.

This changes the saved file, mainly by causing it to have a 2-space
indent instead of 4 (orjson only supports omitting indentation and
using a 2-space indent), but also in that some characters that the
standard library json module emits escaped in strings are instead
present literally in strings.