Set 1 minute HTTP request timeout for completions
Instead of the OpenAI library's default of 10 minutes.

This also moves a logging call from one function to another to
improve reporting (because the function's name is included in the
output, and there is currently no need to do otherwise), changing
the message accordingly to avoid repeating information.

The real primary reason for that logging change is to prepare for
running multiple completion requests at the same time, since the
speed in the test appeared much lower than the rate limit. But this
commit does *not* add that functionalty.