Add completions for definitions
This run successfully generated a completion associated with each
algorithm (or data structure) name, prompted to give a definition.
These may or may not be the final completions for "Experiment #2",
but I am going to embed them and check the similarities, to see
how well whatalgo seems to perform when model-generated definitions
are used rater than the algorithm names themselves. I *might*
subsequently vary the completion model, prompt, and/or inference
parameters to see if I can improve it.

The completions are included, in whatalgo-definitions.json.

[NOTE (added during rebasing): I was, fortunately, confused, in
what I wrote below. The total number of tokens across all texts in
a single request is not constrained to anything like the context
length (max tokens) of the model. It can be enormously greater than
8191 (so long as each individual text being embedded with
text-embedding-ada-002 is <=8191 tokens). In particular, the use
case here does not require breaking them up into smaller batches.
Subsequent runs have succeeded with no error. The text below is
retained for posterity.]

Although generating the completions succeeded, the notebook shows
an error attempting to embed them. I expected an error, since there
are 940 algorithm names, and I believe text-embedding-ada-002's
token limit applies to each *request*, even if it is a request to
embed multiple separate texts each of which is <= 8191 tokens: the
mean length of the definitions would have to be under 9 tokens for
that to work! However, the actual error shown in the notebook
appears different: it is an HTTP 500 error, raising an APIError
exception (https://help.openai.com/en/articles/6897179-apierror).

That happens sometimes; my guess is that when the request is
retried, it will fail with an error specific to the combined
length of the texts. (I haven't done that yet, because I want the
notebook committed with the exact contents it showed after getting
the completions.)