Regenerate definitions and compute similarities
The actual embeddings are not committed here. I plan to add them
later, or to add equivalent or better embeddings depending on what
else can be improved. (Committing many versions of 1536-dimensional
embeddings of ~940 texts will make the repository grow faster than
I would like, and if this is avoided when not clearly helpful, then
both slow fetching and the need to use Git LFS might be avoided.)

So... Good news first: my expectation in 5261227, that something
more involved than a single call to embed_many would be needed to
embed texts whose total length amounted to far more than 8191
tokens, was fortunately a misunderstanding! This worked fine.

Now the bad news: On the (admittedly limited) test code inputs to
classify, whatalgo also fails when the feature extractor function
(passed as "name_describer") is the definition-prompted completion
generating function "define_names" (experiment #2) rather than the
identity function "same_names" (experiment #1).

There are a few major approaches I could try, such as using a
completion model to generate pseudocode, or definitions with
pseudocode, or short characterizing blurbs with short
implementations in multiple programming languages, or just short
Python programs, or programs/pseudocode with all variables and
functions names with a single letter.

The potential benefit of the latter is because the problem is
likely related, a least partly, to misleading overlap between
fragments of English in the implementations being
classified--comments, docstrings, and variable and function
names--to words in the ~940 "known algorithms"' names or
model-generated descriptions.

In that vein, there are approaches in how the data are prepared
before being embedded, and in how the model's output can be used,
that I haven't tried. I'd prefer not to rely on the input being
Python, and I'd definitely prefer not to rely on it being well
formed (though the test inputs I wrote are well-formed Python). But
there may be some reasonable way to heuristically remove comments
and docstrings that doesn't lock it into a single language.

The problem with that--and also with having the feature extractor
for the "known names" produce output with meaningless variable and
function names--is that my goal is to be able to classify code
*even if* it doesn't have expected comments, docstrings, function
or variable names, etc. But much of the time, these are as
expected, and I'd prefer to have reasonably good performance then
too. Of course, this can be relaxed somewhat since the goal here is
as a demonstration rather than to have this deployed large scale
without modification! But I'd still prefer not to hurt common use
cases without a clear good reason.

Another approach, if I am right that this is the main source of the
poor performance in classification (or if I am wrong but I am able
to find out the source and it is likewise systematic) could be to
correct for the effect. For example, I might be able to find a way
to automatically adjust the similarities to account for the
undesired effect of word similarity.

One way to do this could be to also use less "advanced" language
processing techniques, such as judging the similarity based on
natural-language words (or even just English words) using a
technique like TF-IDF on lemmatizable substrings consisting of
letters, or using spaCy word vectors on the whole code and
descriptions. Then, assuming this performs even worse (which I
expect), it can be used *in reverse* to adjust model output.
Alternatively, two separate classifiers for the "known names" could
be used, neither adequately good but one worse, and if it is worse
in the way they are both systematically bad, then the worse one
could be used indicate the direction opposite the adjustment in
similarities that should be done.

Of course, before I do any of that, I should probably try the more
modest change of varying the completion model (I don't plan to try
GPT-4 for this, since it too expensive to be a good choice for
generating this volume of text, but I still have a reasonable
choice of multiple models) and the completion model's inference
parameters (such as temperature) and prompt.