Using 20 threads seems fastest
This sets the number of threads used to make concurrent requests to
the completion model to 20. This number seems faster than 15 and
25. Using 20 threads took about 8 minutes, whereas 15 threads had
taken about 10 minutes. (25 threads had exceeded the rate limits a
number of times, which is unlikely to be good for performance, and
I cancelled the run, so I don't have a specific timing for that.)

Of course, it might be worth it to use asynchronous programming
for the completions (asyncio and openai.ChatCompletion.acreate
instead of openai.ChatCompletion.create). I'm not speaking to
which broad approach is better here, just for what a good number of
threads seems to be when using ThreadPoolExecutor.

Since this probably depends in significant part on network
conditions and load on OpenAI's servers, it may be better not to
use a fixed number of concurrent connections. But the techniques
for that are more complicated.

As before, I'm not committing the new embeddings, since this run of
experiment #2 is not valuable for its results, which are in no
important way different from previous results. However, in this
case I am also not committing the generated definitions, just the
small change in the module and the notebook changes showing the
run.