Alternative definition prompt, different wrong matches
The idea here is to get definitions that are worded in a way that
omits some otherwise common words in the jargon of algorithms that
may appear in comments, docstrings, and variable and function names
in code inputs being classified, since that appears central to the
problems finding correct matches for the components_{a,b,c} inputs
I've been testing with.

Although this does achieve that, sort of, I think it ends up using
other such words more often, so it doesn't really do better.

I've omitted not just the embeddings but also the generations, to
save space. (Some examples are viewable in the notebook, though.)