Tweak prompt for generating pseudocode
This also doesn't work. As before, I'm committing it to keep a
record.

As for the prompt, I have no idea what "adiabatic pseudocode" is.
There are adiabatic thermal processes, adiabatic quantum computing,
and the word "adiabatic" applies in simulated annealing and related
algorithms, but none of that seems to explain the effect of this
prompt. "Adiabatic" means "with no loss or gain of heat or mass."

I got this prompt by experimenting with a whole bunch of variations
(in the OpenAI Playground), including qualifying "pseudocode" with
a variety of adjectives, including those that seem inapplicable or
absurd. For some reason, adding "adiabatic" seems to make it follow
the prompt better: omiting text outside the pseudocode, omitting
backticks that would indicate a code fence, omitting comments,
writing the pseudocode in a more code-like style, and writing the
pseudocode in a somewhat more uniform style across separate
queries. Most adjectives don't seem to have any such effect, but
it seems that "pyloric" has a similar effect. ("Pyloric" means
"relating to the connection between the stomach and the small
intestine.")

Although the prompt tweak appears to have some beneficial effects
in the prompts I've inspected, I do not claim that it causes an
overall increase in readability, accuracy, or educational value.
However, those are not primary goals in the generations of these
prompts, which are significant for their embeddings in the search
problem. But they don't work better than the non-"adibatic" ones
from before, either, unfortunately.

As before, I've omitted both the embeddings and the generations
from this commit, to save space.