Give Timeout 6 max tries too (in whatalgo)
Instead of 10.

Besides reducing the number of successive timeout failures needed
to give up, this is also significant in that the maximum number of
interleaved failures of different kinds that can occur before
giving up is lower. This is due to the way decorators retry their
wrapped functions up to a specific number of times, but when the
wrapped function itself is retrying, the effect can be
multiplicative if the failure types are interleaved just right.