Temporarily make CI tests non-fail-fast
This was originally only for the poetry tests, and was done to
investigate the failures blocking #311, observable in various runs
where the `test` workflow is enabled, EliahKagan#85, in:
https://github.com/EliahKagan/EmbeddingScratchwork/actions/runs/15377782244/job/43264696136?pr=85

The failure occurred in at least one macOS job. The direct cause
was that there is no Fortran compiler available to build code in
`scipy` that requires it. But the root cause was that a version, or
specific wheels of it, were wrongly or non-optimally resolved on at
least some versions and architectures, so that pre-built binaries
were not being used, even though such binaries had been published.
(Ordinarily, native code compilation is not required for any of our
dependencies, not even indirect ones.)

This remains relevant because it still occurs in most or all
Dependabot version update and Dependabot security update PRs for
Python packages, requiring that updates be done more manually
until fixed (or until some alternative to Dependabot is used).

This now also makes the conda CI tests non-fail-fast, to
investigate a seemingly separate HTTP 502 Bad Gateway failure that
occurs on the main branch in the `test` workflow, such as in:
https://github.com/EliahKagan/EmbeddingScratchwork/actions/runs/16009371598/job/45164849861