Temporarily make CI poetry tests non-fail-fast
This is to investigate the failures blocking #311, which are
observable in EliahKagan#85, in:
https://github.com/EliahKagan/EmbeddingScratchwork/actions/runs/15377782244/job/43264696136?pr=85

The failure occurs in at least one macOS job, and the direct cause
seems to be that there is no Fortran compiler available to build
code in `scipy` that requires it. But the root cause may be that a
version is (wrongly?) being selected on at least some versions and
architectures of macOS that doesn't have published pre-build
binaries. Ordinarily, I don't think any native code compilation is
required for any of our (even indirect) dependencies.